1-ingestao-e-busca-semantica/requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-postgres
langchain-text-splitters
pypdf
psycopg[binary]
python-dotenv
